+++ tests/mocks/mock_multi_level_sensor_device.py
@@ -11,19 +11,18 @@ from devolo_home_control_api.properties.settings_property import SettingsPropert
 from .mock_gateway import MockGateway
 
 
-def multi_level_sensor_device(key: str) -> Zwave:
+def multi_level_sensor_device(device_uid: str) -> Zwave:
     """
     Represent a Multi Level Sensor
 
-    :param key: Key to look up in test_data.json
+    :param device_uid: Device UID this mock shall have
     :return: Multi Level Sensor device
     """
     file = pathlib.Path(__file__).parent / ".." / "test_data.json"
     with file.open("r") as fh:
         test_data = json.load(fh)
 
-    device = Zwave(**test_data.get("devices").get(key))
-    device_uid = test_data.get("devices").get(key).get("uid")
+    device = Zwave(**test_data.get("devices").get("sensor"))
     gateway = MockGateway(test_data.get("gateway").get("id"))
     session = requests.Session()
 
@@ -36,25 +35,25 @@ def multi_level_sensor_device(key: str) -> Zwave:
         BinarySensorProperty(gateway=gateway,
                              session=session,
                              element_uid=element_uid,
-                             state=test_data.get("devices").get(key).get("state"))
+                             state=test_data.get("devices").get("sensor").get("state"))
 
     element_uid = f'devolo.MultiLevelSensor:{device_uid}#MultilevelSensor(1)'
     device.multi_level_sensor_property[element_uid] = \
         MultiLevelSensorProperty(gateway=gateway,
                                  session=session,
                                  element_uid=element_uid,
-                                 value=test_data.get("devices").get(key).get("value"),
-                                 unit=test_data.get("devices").get(key).get("unit"))
+                                 value=test_data.get("devices").get("sensor").get("value"),
+                                 unit=test_data.get("devices").get("sensor").get("unit"))
 
     device.settings_property['temperature_report'] = SettingsProperty(gateway=gateway,
                                                                       session=session,
                                                                       element_uid=f"trs.{device_uid}",
-                                                                      temp_report=test_data.get("devices").get(key).get(
+                                                                      temp_report=test_data.get("devices").get("sensor").get(
                                                                           "temp_report"))
     device.settings_property['motion_sensitivity'] = SettingsProperty(gateway=gateway,
                                                                       session=session,
                                                                       element_uid=f"mss.{device_uid}",
-                                                                      motion_sensitivity=test_data.get("devices").get(key).get(
+                                                                      motion_sensitivity=test_data.get("devices").get("sensor").get(
                                                                           "motion_sensitivity"))
 
     return device